% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{breaks_integer}
\alias{breaks_integer}
\title{Integer breaks for continuous axes}
\usage{
breaks_integer(x, n = 5, ...)
}
\arguments{
\item{x}{(numeric) Data mapped onto continuous axis}

\item{n}{(numeric) Desired number of breaks. Passed onto `pretty()`}

\item{...}{Other arguments passed on to `pretty()`}
}
\description{
Return a vector of integer breaks for a numeric axis.
This can cause breaks to have irregular intervals.
}
\examples{
x <- seq(from = 0.5, to = 3.8, by = 0.15)
breaks_integer(x)
}
