% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{breaks_interval}
\alias{breaks_interval}
\title{Breaks at specified intervals}
\usage{
breaks_interval(x, i = NULL, ...)
}
\arguments{
\item{x}{(numeric) Data mapped onto continuous axis}

\item{i}{(numeric) Interval between breaks}

\item{...}{Other arguments passed on to `labeling::extended()`}
}
\description{
Return a vector of breaks for a numeric axis.
Based on `scales::breaks_extended()`, but will observe the specified interval
}
\examples{
x <- seq(1, 128)
breaks_interval(x, i = 40)
}
