% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{deframe_with_class}
\alias{deframe_with_class}
\title{Deframe and set class}
\usage{
deframe_with_class(dat, class)
}
\arguments{
\item{dat}{(data.frame) Table with one or two columns}

\item{class}{(character) Class to which to set to elements of resulting vector}
}
\value{
A vector
}
\description{
Deframe tibble into named vector and set class of each element
}
\examples{
dat <- tibble::tibble(var = c("a", "b", "c"), value = c(1, 2, 3))
deframe_with_class(dat, class = "integer")
}
\seealso{
[glebrt::set_class()] [tibble::deframe()]
}
