% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{deselect_na}
\alias{deselect_na}
\title{Deselect NA columns}
\usage{
deselect_na(dat, keep = NULL)
}
\arguments{
\item{dat}{Data frame}

\item{keep}{Names of columns to keep despite them being entirely `NA`}
}
\description{
Remove NA columns from a data frame while allowing to retain specific columns
in the same position
}
\examples{
dat <- tibble::tibble(
  var = letters[1:5],
  value1 = 1:5,
  value2 = rep(NA, 5)
)
}
