% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerate.R
\name{enumerate}
\alias{enumerate}
\title{Count unique values in column}
\usage{
enumerate(dat, column)
}
\arguments{
\item{dat}{(data.frame)}

\item{column}{(character) Name of column to count unique values in}
}
\value{
An integer
}
\description{
Get the number of unique values from a column of a data.frame
}
\examples{
dat <- tibble::tibble(var = c("a", "b", "c", "b"))
enumerate(dat, var)
}
