% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{get_greys}
\alias{get_greys}
\title{Return a Scale of Greys}
\usage{
get_greys(i = NULL)
}
\arguments{
\item{i}{(numeric) Vector of indices of greys to return}
}
\value{
A character vector of colour names
}
\description{
Returns up to five different shades of grey
}
\examples{
get_greys(1)
get_greys(1:3)
get_greys(c(1, 3, 5))
}
