% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_precision.R
\name{get_precision}
\alias{get_precision}
\title{Get Precision of Number}
\usage{
get_precision(x)
}
\arguments{
\item{x}{(numeric) Vector to get precision of}
}
\description{
Get the numeric precision of numbers (i.e. the number of significant digits)
}
\examples{
x <- c(1, 0.4, 0.001, 0.23, 0.170)
get_precision(x)
}
