% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glebrt-package.R
\docType{package}
\name{glebrt-package}
\alias{glebrt}
\alias{glebrt-package}
\title{glebrt: Gleb Ebert's Commonly Used Functions}
\description{
glebrt contains a collection of functions that I commonly use in my R projects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gl-eb/glebrt}
  \item Report bugs at \url{https://github.com/gl-eb/glebrt/issues}
}

}
\author{
\strong{Maintainer}: Gleb G. Ebert \email{dev@gl-eb.me} (\href{https://orcid.org/0000-0002-7734-3824}{ORCID}) [copyright holder]

}
\keyword{internal}
