% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_conjuction.R
\name{paste_conjunction}
\alias{paste_conjunction}
\title{Collapse vector with conjunction}
\usage{
paste_conjunction(x, conjunction = "and")
}
\arguments{
\item{x}{Object that both [paste0()] and [length()] can be called on}

\item{conjunction}{(character) The conjunction to be used behind the oxford comma}
}
\value{
String
}
\description{
Paste together vector with final conjunction, "and" for example, as well as an oxford comma
}
\examples{
paste_conjunction(c("a", "b", "c"))
paste_conjunction(c("a", "b", "c"), conjunction = "as well as")
}
