% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision.R
\name{ceiling_precision}
\alias{ceiling_precision}
\alias{floor_precision}
\title{Rounding of Numbers With Precision}
\usage{
ceiling_precision(x, precision = 2L)

floor_precision(x, precision = 2L)
}
\arguments{
\item{x}{(numeric) Vector to round}

\item{precision}{(numeric) Integer indicating the precision to which to round to. Non-integer numerics will be passed to `as.integer()`.}
}
\value{
A numeric vector of the same length as `x`
}
\description{
`ceiling_precision` rounds up the nearest number of specified decimal places.
`floor_precision` rounds down the nearest number of specified decimal places.
}
\examples{
x <- c(0.001, 0.23, 0.170)
ceiling_precision(x, precision = 2)
floor_precision(x, precision = 2)
}
