% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{set_class}
\alias{set_class}
\title{Set class}
\usage{
set_class(x, class)
}
\arguments{
\item{x}{Object to which to apply the class to}

\item{class}{(character) Class to which to set to elements of resulting vector}
}
\description{
A function for setting classes inside of pipes
}
\examples{
set_class(1, class = "integer")
}
