% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_ggplot_theme.R
\name{set_ggplot_theme}
\alias{set_ggplot_theme}
\title{Set a global ggplot2 theme}
\usage{
set_ggplot_theme(
  theme = "theme_light",
  base_size = 15,
  font_family = "Fira Sans",
  ink = "black",
  paper = "white",
  accent = "#C83F4C"
)
}
\arguments{
\item{theme}{(character) The name of a ggplot2 complete theme (default: "theme_light")}

\item{base_size}{(numeric) Base font size argument (base_size) passed onto ggplot theme}

\item{font_family}{(character) Font family name}

\item{ink, paper, accent}{(character) colour for foreground, background, and accented elements respectively.}
}
\description{
This function sets a global ggplot2 theme using ggplot2::set_theme().
In case you want to use non-default wont weights, you need to register them
as a font variant first (see `systemfonts::register_variant()`).
}
\details{
Arguments are eventually passed on to the theme function as defined
in `theme`. For font selection to work, you need to use a graphics device
with builtin `systemfonts` support like `ragg` or `svglite`.
}
\examples{
\dontrun{
set_ggplot_theme()
set_ggplot_theme(theme = "theme_light")
set_ggplot_theme(
  theme = "theme_light",
  base_size = 15,
  font_family = "Fira Sans",
)
}
}
\seealso{
[ggplot2::ggtheme], [systemfonts::register_variant()]
}
