% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{some_of}
\alias{some_of}
\alias{same}
\title{Tests for logical vectors}
\usage{
some_of(x)

same(x)
}
\arguments{
\item{x}{(logical) Vector to be tested}
}
\description{
`some_of()` checks if some but not all elements of a logical vector are TRUE.

`same()` checks if all elements of a logical vector are the same.
}
\examples{
some_of(c(TRUE, FALSE, FALSE))
some_of(c(TRUE, TRUE, TRUE))
some_of(c(FALSE, FALSE, FALSE))

same(c(TRUE, FALSE, FALSE))
same(c(TRUE, TRUE, TRUE))
same(c(FALSE, FALSE, FALSE))
}
