% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_text.R
\name{spell_numbers}
\alias{spell_numbers}
\title{Spell out numbers}
\usage{
spell_numbers(x, conjunction = NULL)
}
\arguments{
\item{x}{Numeric vector to be converted to text}

\item{conjunction}{Conjunction to be used behind the oxford comma. Usually a character vector vector of length one, but can be any object that can be converted to a character vector}
}
\value{
Character vector of length 1
}
\description{
Spell out and concatenate a numeric vector if any of the numbers are >10.
Vectors of length >1 are concatenated with an optional conjunction.
`spell_numbers()` wraps `xfun::numbers_to_words()` and adds some additional logic.
}
\examples{
spell_numbers(2)
spell_numbers(20)
spell_numbers(c(1, 2, 3))
spell_numbers(c(1, 2, 30))
spell_numbers(c(1, 2, 3), conjunction = "and")
spell_numbers(c(1, 2, 30), conjunction = "and")
}
