% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_text.R
\name{spell_range}
\alias{spell_range}
\title{Spell out range}
\usage{
spell_range(x)
}
\arguments{
\item{x}{Numeric vector to be converted to text}
}
\value{
Character vector of length 1
}
\description{
Spell out the range of a numeric vector if any of the numbers are >10.
`spell_range()` uses `xfun::numbers_to_words()` for number to text conversion.
}
\examples{
spell_range(c(1, 2))
spell_range(c(1, 2, 3))
spell_range(c(1, 2, 3, 30))
spell_range(c(20, 25, 30))
}
