% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_objects.R
\name{unset_legend}
\alias{unset_legend}
\title{Reset ggplot2 legend and guides to defaults}
\usage{
unset_legend(gg)
}
\arguments{
\item{gg}{ggplot2 object as returned by `ggplot()`}
}
\description{
`unset_legend()` takes a ggplot2 object, unsets legend components of its
theme, and resets the guides to default values. This can be useful when
combining multiple plots into a figure with patchwork, which requires guides
and legends to be exactly the same when collecting them across panels.
}
\examples{
dat <- tibble::tibble(
  var = seq(1:10),
  val = rev(seq(1:10))
)
plot <- dat |>
  ggplot2::ggplot(ggplot2::aes(x = var, y = val, colour = val)) +
  ggplot2::geom_point() +
  ggplot2::scale_colour_continuous(breaks = dat$val) +
  ggplot2::guides(
    colour = ggplot2::guide_legend()
  ) +
  ggplot2::theme(
    legend.position = "bottom"
  )

# unmodified plot
plot

# plot with legend and guide modifications restored to defaults
unset_legend(plot)
}
