% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_text.R
\name{paste_conjunction}
\alias{paste_conjunction}
\title{Collapse vector with conjunction}
\usage{
paste_conjunction(x, conjunction = "and")
}
\arguments{
\item{x}{Usually a character vector, but can be any object that can be converted to a character vector and is subsettable}

\item{conjunction}{Conjunction to be used behind the oxford comma. Usually a character vector vector of length one, but can be any object that can be converted to a character vector}
}
\value{
Character vector of length 1
}
\description{
Paste together vector with final conjunction like "and", as well as an oxford comma
}
\examples{
paste_conjunction(c("a", "b", "c"))
paste_conjunction(c("a", "b", "c"), conjunction = "as well as")
}
