% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_ggplot_theme.R
\name{set_ggplot_theme}
\alias{set_ggplot_theme}
\title{Set a global ggplot2 theme}
\usage{
set_ggplot_theme(
  theme = "theme_light",
  base_size = 15,
  font_family = list(font_name = "Fira Sans", font_filename = "FiraSans", font_format =
    "otf", font_cut_roman = "Regular", font_cut_italic = "Italic", font_weight_bold =
    "Medium")
)
}
\arguments{
\item{theme}{(character) The name of a ggplot complete theme (default: "theme_light")}

\item{base_size}{(numeric) Base font size argument (base_size) passed onto ggplot theme}

\item{font_family}{(character or list) Font family name or font family list. See examples}
}
\description{
This function sets a global ggplot2 theme using ggplot2::theme_set()
}
\examples{
set_ggplot_theme()
set_ggplot_theme(theme = "theme_light")
\dontrun{
set_ggplot_theme(
  theme = "theme_light",
  base_size = 15,
  font_family = list(
    font_name = "Fira Sans",
    font_filename = "FiraSans",
    font_format = "otf",
    font_cut_roman = "Regular",
    font_cut_italic = "Italic",
    font_weight_bold = "Medium"
  )
)
}
}
