% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_ggplot_theme.R
\name{set_ggplot_theme}
\alias{set_ggplot_theme}
\title{Set a global ggplot2 theme}
\usage{
set_ggplot_theme(
  theme = "theme_light",
  base_size = 15,
  font_family = "Fira Sans"
)
}
\arguments{
\item{theme}{(character) The name of a ggplot2 complete theme (default: "theme_light")}

\item{base_size}{(numeric) Base font size argument (base_size) passed onto ggplot theme}

\item{font_family}{(character) Font family name}
}
\description{
This function sets a global ggplot2 theme using ggplot2::set_theme().
In case you want to use non-default wont weights, you need to register them
as a font variant first (see `systemfonts::register_variant()`).
}
\examples{
\dontrun{
set_ggplot_theme()
set_ggplot_theme(theme = "theme_light")
set_ggplot_theme(
  theme = "theme_light",
  base_size = 15,
  font_family = "Fira Sans",
)
}
}
